Moved new types to production target: `RemoteImageCommentAuthor`, `RemoteImageComment`, `ImageCommentsMapper`, `RemoteCommentsLoader`
