+++ EssentialFeed/EssentialFeed/Image Comments API/RemoteImageCommentAuthor.swift
@@ -1,0 +1,13 @@
+//
+//  RemoteImageCommentAuthor.swift
+//  EssentialFeed
+//
+//  Created by Bogdan Poplauschi on 28/03/2021.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+struct RemoteImageCommentAuthor: Decodable {
+	let username: String
+}
