+++ EssentialFeed/EssentialFeedTests/Image Comments API/LoadImageCommentsFromRemoteUseCaseTests.swift
@@ -9,91 +9,6 @@
 import EssentialFeed
 import XCTest
 
-final class RemoteCommentsLoader {
-	
-	enum Error: Swift.Error {
-		case connectivity
-		case invalidData
-	}
-	
-	typealias Result = Swift.Result<[ImageComment], Swift.Error>
-	
-	private let url: URL
-	private let client: HTTPClient
-	
-	init(url: URL, client: HTTPClient) {
-		self.url = url
-		self.client = client
-	}
-	
-	func load(completion: @escaping (Result) -> Void) {
-		client.get(from: url) { [weak self] result in
-			guard self != nil else { return }
-			
-			switch result {
-			case .failure:
-				completion(.failure(Error.connectivity))
-				
-			case let .success((data, response)):
-				completion(RemoteCommentsLoader.map(data, from: response))
-			}
-		}
-	}
-	
-	private static func map(_ data: Data, from response: HTTPURLResponse) -> Result {
-		do {
-			let imageComments = try ImageCommentsMapper.map(data, from: response)
-			return .success(imageComments.toModels())
-		} catch {
-			return .failure(error)
-		}
-	}
-}
-
-private final class ImageCommentsMapper {
-	private struct Root: Decodable {
-		let items: [RemoteImageComment]
-	}
-	
-	static func map(_ data: Data, from response: HTTPURLResponse) throws -> [RemoteImageComment] {
-		guard response.statusCode == 200 else {
-			throw RemoteCommentsLoader.Error.invalidData
-		}
-		
-		do {
-			let jsonDecoder = JSONDecoder()
-			jsonDecoder.dateDecodingStrategy = .iso8601
-			let root = try jsonDecoder.decode(Root.self, from: data)
-			return root.items
-		} catch {
-			throw RemoteCommentsLoader.Error.invalidData
-		}
-	}
-}
-
-private extension Array where Element == RemoteImageComment {
-	func toModels() -> [ImageComment] {
-		return map {
-			ImageComment(
-				id: $0.id,
-				message: $0.message,
-				createdAt: $0.created_at,
-				author: ImageCommentAuthor(username: $0.author.username))
-		}
-	}
-}
-
-private struct RemoteImageComment: Decodable {
-	let id: UUID
-	let message: String
-	let created_at: Date
-	let author: RemoteImageCommentAuthor
-}
-
-private struct RemoteImageCommentAuthor: Decodable {
-	let username: String
-}
-
 final class LoadImageCommentsFromRemoteUseCaseTests: XCTestCase {
 	func test_init_doesNotRequestDataFromURL() {
 		let (_, client) = makeSUT()
